refactor  - space damafge is indexed by object ID  - ship manager space object is `DeepReadonly<Spaceship>`
