+++ package.json
@@ -79,6 +79,7 @@
         "prettier": "^2.6.2",
         "pretty-quick": "^3.1.0",
         "rimraf": "^3.0.2",
+        "ts-essentials": "^9.1.2",
         "ts-jest": "^27.1.4",
         "ts-loader": "^9.3.0",
         "ts-node": "^10.7.0",
